Split messages into lawyers and law_firms routes
